+++ lib/grape/endpoint.rb
@@ -1,7 +1,3 @@
-require 'rack'
-require 'grape'
-require 'hashie'
-
 module Grape
   # An Endpoint is the proxy scope in which all routing
   # blocks are executed. In other words, any methods