+++ lib/grape/entity.rb
@@ -1,5 +1,3 @@
-require 'hashie'
-
 module Grape
   # An Entity is a lightweight structure that allows you to easily
   # represent data from your application in a consistent and abstracted
@@ -333,6 +331,11 @@ module Grape
     def to_json(options = {})
       options = options.to_h if options && options.respond_to?(:to_h)
       MultiJson.dump(serializable_hash(options))
+    end
+
+    def to_xml(options = {})
+      options = options.to_h if options && options.respond_to?(:to_h)
+      serializable_hash(options).to_xml(options)
     end
 
     protected