+++ lib/grape/validations.rb
@@ -1,7 +1,3 @@
-require 'virtus'
-require 'i18n'
-
-I18n.load_path << File.expand_path('../locale/en.yml', __FILE__)
 module Grape
 
   module Validations