+++ spec/spec_helper.rb
@@ -12,10 +12,7 @@ Bundler.setup :default, :test
 
 require 'rack/test'
 require 'pry'
-
 require 'base64'
-
-require 'hashie/hash'
 
 Dir["#{File.dirname(__FILE__)}/support/*.rb"].each do |file|
   require file